Begin RecommendedMeal component, a random meal
The RecommendedMeal component has logic that determines the time of day of the user (for now it does a silly, trivial check -- we will make it better later) and recommends an appropriate meal based on the time. It shows basic information, such as a picture of the meal and its name. It is clickable, and doing so sends the user off to a detailed page showing more about that meal.