Refactor the way Page components are imported.
This practice of including all Components related to a single "Feature" (in this case, a "Page") is pretty standard. We use an index file to import all the Page components, and then export them as a single object. This then allows us, for example, in App.js to import all Page components with a single import statement.