Create helpers for mock data, meal, & time
Created a helper folder that will store helper files -- each file serving to house closely-related features together. The mockData file contains helpers that let us quickly work with a fake User and Meal object. The time file contains helpers for anything that needs the user's time -- such as if it's breakfast, lunch, or dessert time. Finally, the meal file contains helpers for anything related to a meal, such as the recommended meal category, the greeting message, and the recommendation message. (Admittedly 'meal' might be a bad category of helpers, we can reconsider this and all others.)