Remove time functions. Add more styling.
I refactored the functions that determine the user's time and calculate some result (such as which recipe type to recommended, and what message to show) and moved them into an appropriate helper file. I developed the styling of the RecommendedMeal component a little more, but it still has a lot of work to be done!