More cateogries. Can fetch results by category.
I added the other categories that the Spoonacular API allows to be used for querying. Some of them I did not find a suitable emoji for, yet. Clicking the 'See all' button increases the number of categories shown from the initial 5 to the full size of 14. The initial category is chosen based off the user's time (Breakfast, Main Course, or Dessert). Clicking a category retrieves the first 10 results. You can then click on one of the results and be taken to the detailed view of that recipe. A lot of API and Styling work to do! This was just to get the ball rolling.