Add very minor styling
Began revamping the styling of the Recipe page to look more like the Dribbble inspiration. Couldn't take it too much further, as it looks like we ran out of API calls for the day. I'll have to create some more dummy meal data properties (description, ingredients, etc) so I have some data to look at when working on this screen!