Add CORS middleware. Add fake user data to /users
Need the cors middleware to get rid of that annoying "Blocked due to CORS policy" when making a request to our server from our React client. Added some hard-coded users data to the /users controller, so we can test calling this route from our React client. Shortly, this hard-coded data will be stored in a database!