Begin logic for Search Page
By visiting '/search/<searchTerm>', the app will appropriately fetch 5 recipes that have that search term as a match. For now, there is no input for the user to type their own search -- we can only enter the search directly into the /search/ URL.

The styling of the returned list of recipes is very basic and unfinished. But you can click on a result and be taken to its detailed recipe page!