Make User schema store favoriteMeals as references
Previously, favoriteMeals was an array of Strings (each String representing the ID of a Meal). Now, it is an array of mongoose.Schema.Types.ObjectId. This will allow us to store these meals as references to actual Meal schemas. Consequently, in userController we populate favoriteMeals with the actual Meal model!