Fix Link bug by using Stack rather than Flex
The MealPreview's anchor tag was clickable way outside of its physical boundaries, due to Chakra's very odd implementation of Flex. Instead, I chose to use a Stack element, because it magically fixes it. What is a Stack? Why did it make the React Link behave so oddly? Who cares -- this is Chakra!