Create Skeleton version of MealPreview
This is a temporary component where I experimented with how to style various Chakra Skeleton elements to work with our MealPreview component. In the end, we would most likely just directly incorporate this logic into the MealPreview component itself, and not have a separate one. But I did not want to mess up MealPreview as I tested stuff