Use Link instead of useNavigate. Tweak animation
Rather than using useNavigate to direct the user to the user's profile, I used a Link instead. I also learned how Chakra can apply transition to a child when the parent is hovered, so I modified the previous animation (scaling the entire greeting message up when hovered) to just change the color of the user's name when hovered.