Improve alignment, spacing, and padding
I modified the Meal Preview cards to assume the title of the meal is going to be two lines (which I also enforce by truncating anything over 2 lines). I also decreased the spacing between the title, difficulty, and time to prepare. Padding was also added to the bottom of the card. Finally, I added a small border around the image, to make them stand out more. Like with other components today, I also removed Stack and used a Flex instead, as it does not have that weird alignment issue.