Improve font hierarchy, tweak border
I improved the font hierarchy between the various Text elements by scaling down the time to prepare / difficulty, increasing the font size of the meal title (and making it bold) and slightly increasing the font size for the header. I also changed the border to be thicker and use our 'secondary' gray color.